make a special dataloader as requested
