+++ utilities/data.py
@@ -1,47 +1,46 @@
+import csv
 import os
 from typing import Optional
 
 import torch
-from lightning.pytorch import LightningDataModule
 from torch.utils.data import DataLoader, Dataset
 from transformers import PreTrainedTokenizer
 
 
-class TextClassificationDataModule(LightningDataModule):
+class TextDataset(Dataset):
+    def __init__(self, csv_file):
+        super().__init__()
+        with open(csv_file, newline="") as csvfile:
+            reader = csv.reader(csvfile, delimiter=",")
+            self.rows = list(reader)
+
+    def __len__(self):
+        return len(self.rows)
+
+    def __getitem__(self, item):
+        label, text = self.rows[item]
+        label = int(label) - 1
+        return dict(text=text, label=label)
+
+
+class TextClassificationDataLoader(DataLoader):
     def __init__(
         self,
-        train_dataset: Dataset,
-        val_dataset: Dataset,
+        dataset: Dataset,
         tokenizer: PreTrainedTokenizer,
-        batch_size: int = 4,
         max_token_len: int = 256,
+        batch_size: int = 4,
         num_workers: Optional[int] = None,
+        **kwargs,
     ):
-
-        super().__init__()
-        self.train_dataset = train_dataset
-        self.val_dataset = val_dataset
-        self.batch_size = batch_size
-        self.tokenizer = tokenizer
-        self.max_token_len = max_token_len
-        self.num_workers = num_workers if num_workers is not None else os.cpu_count()
-
-    def train_dataloader(self):
-        return DataLoader(
-            self.train_dataset,
-            batch_size=self.batch_size,
-            shuffle=False,
-            num_workers=self.num_workers,
-            collate_fn=TextEncodingCollate(self.tokenizer, self.max_token_len),
-        )
-
-    def val_dataloader(self):
-        return DataLoader(
-            self.val_dataset,
-            batch_size=self.batch_size,
-            shuffle=False,
-            num_workers=self.num_workers,
-            collate_fn=TextEncodingCollate(self.tokenizer, self.max_token_len),
+        collate_fn = TextEncodingCollate(tokenizer, max_token_len)
+        num_workers = num_workers if num_workers is not None else os.cpu_count()
+        super().__init__(
+            dataset,
+            batch_size=batch_size,
+            num_workers=num_workers,
+            collate_fn=collate_fn,
+            **kwargs
         )
